Update swarm system scripts
Create a bash script that automatically sets up a
swarm system of NCP using GlusterFS to replicate data
of all worker nodes and reserve data persistence when
the stack is deleted and created again.

Status: Docker-machines are only using private IPs
(HostOnly in Virtual Box terms) and overlay network
seems to have conflict with this. Therefore, we will
be using Vagrant VMs to represent nodes.